Integra MesaSessao e Pagamento
- Adiciona lista de Pedidos na entidade MesaSessao para calcular o valor total com taxa de serviço.
- Atualiza serviço MesaSessaoService para incluir taxa de serviço no valor total da mesa.
- Altera todas as variáveis do tipo Double para BigDecimal para maior precisão em cálculos financeiros.
- Refatora DTOs relacionados para refletir mudanças na estrutura de dados.
- Garante que o valor total da mesa inclua a taxa de serviço ao fechar a mesa.

+++ TCC-Backend/src/main/java/com/example/restapp/dto/PedidoDTO/PedidoResponseDTO.java
@@ -1,5 +1,6 @@
 package com.example.restapp.dto.PedidoDTO;
 
+import java.math.BigDecimal;
 import java.time.*;
 import lombok.*;
 
@@ -15,7 +16,7 @@ public class PedidoResponseDTO
     private String nomeProduto;
     private String descricaoPedido;
     private Integer quantidadeProduto;
-    private Double valorUnitario;
-    private Double valorTotal;
+    private BigDecimal valorUnitario;
+    private BigDecimal valorTotal;
     private LocalDateTime horarioLancamento;
 }
