+++ TCC-Backend/src/main/java/com/example/restapp/model/financeiro/MetodoPagamento.java
@@ -1,5 +1,7 @@
 package com.example.restapp.model.financeiro;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.*;
 import jakarta.validation.constraints.*;
 import lombok.*;
@@ -25,6 +27,6 @@ public class MetodoPagamento
     @Enumerated(EnumType.STRING)
     private BandeiraCartao bandeira; // Pode ser null para métodos não-cartão
 
-    private double taxa; // Pode ser null se não aplicável
+    private BigDecimal taxa; // Pode ser null se não aplicável
     
 }
